requirements.txt
requests==2.28.2
pytest==7.3.0
python-dotenv==1.0.0
werkzeug==2.2.3